Implement/copy the inference-model of the tensorflow example for cifar-10 (attempt 5)

+++ RFNN/cifar_tf_model.py
@@ -12,7 +12,7 @@ def _variable_on_cpu(name, shape, initializer):
     Variable Tensor
   """
   with tf.device('/cpu:0'):
-    dtype = tf.float64
+    dtype = tf.float32
     var = tf.get_variable(name, shape, initializer=initializer, dtype=dtype)
   return var
   
@@ -33,7 +33,7 @@ def _variable_with_weight_decay(name, shape, stddev, wd):
   Returns:
     Variable Tensor
   """
-  dtype = tf.float64
+  dtype = tf.float32
   var = _variable_on_cpu(
       name,
       shape,
